+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change `PartialEq` and `Eq` impls for `PublicKey`
+
 ## [0.7.0] - 2021-07-05
 
 ### Changed